feat: remove continue button to the parent and update icon
